+++ state-chain/chains/src/sol/sol_tx_core/transaction.rs
@@ -1,13 +1,14 @@
 use crate::sol::{
 	sol_tx_core::{
-		compile_instructions, program::instruction::InstructionError, short_vec,
-		CompiledInstruction, CompiledKeys, Hash, Instruction, MessageHeader, Pubkey, RawSignature,
+		compile_instructions, short_vec, CompiledInstruction, CompiledKeys, Hash, Instruction,
+		MessageHeader, Pubkey, RawSignature,
 	},
 	SolSignature,
 };
 use codec::{Decode, Encode};
 use scale_info::TypeInfo;
 use serde::{Deserialize, Serialize};
+use sol_prim::errors::TransactionError;
 use sp_std::{vec, vec::Vec};
 
 #[cfg(any(test, feature = "runtime-integration-tests"))]
@@ -41,7 +42,7 @@ pub mod legacy {
 		Encode, Decode, TypeInfo, Serialize, Deserialize, Default, Debug, PartialEq, Eq, Clone,
 	)]
 	#[serde(rename_all = "camelCase")]
-	pub struct Message {
+	pub struct LegacyMessage {
 		/// The message header, identifying signed and read-only `account_keys`.
 		// NOTE: Serialization-related changes must be paired with the direct read at sigverify.
 		pub header: MessageHeader,
@@ -59,7 +60,7 @@ pub mod legacy {
 		pub instructions: Vec<CompiledInstruction>,
 	}
 
-	impl Message {
+	impl LegacyMessage {
 		pub fn new_with_blockhash(
 			instructions: &[Instruction],
 			payer: Option<&Pubkey>,
@@ -147,7 +148,7 @@ pub mod legacy {
 	#[derive(
 		Encode, Decode, TypeInfo, Debug, PartialEq, Default, Eq, Clone, Serialize, Deserialize,
 	)]
-	pub struct Transaction {
+	pub struct LegacyTransaction {
 		/// A set of signatures of a serialized [`Message`], signed by the first
 		/// keys of the `Message`'s [`account_keys`], where the number of signatures
 		/// is equal to [`num_required_signatures`] of the `Message`'s
@@ -161,11 +162,11 @@ pub mod legacy {
 		pub signatures: Vec<SolSignature>,
 
 		/// The message to sign.
-		pub message: Message,
+		pub message: LegacyMessage,
 	}
 
-	impl Transaction {
-		pub fn new_unsigned(message: Message) -> Self {
+	impl LegacyTransaction {
+		pub fn new_unsigned(message: LegacyMessage) -> Self {
 			Self {
 				signatures: vec![
 					SolSignature::default();
@@ -177,7 +178,7 @@ pub mod legacy {
 
 		#[cfg(any(test, feature = "runtime-integration-tests"))]
 		pub fn new_with_payer(instructions: &[Instruction], payer: Option<&Pubkey>) -> Self {
-			let message = Message::new(instructions, payer);
+			let message = LegacyMessage::new(instructions, payer);
 			Self::new_unsigned(message)
 		}
 
@@ -263,7 +264,7 @@ pub mod legacy {
 		}
 
 		/// Return the message containing all data that should be signed.
-		pub fn message(&self) -> &Message {
+		pub fn message(&self) -> &LegacyMessage {
 			&self.message
 		}
 
@@ -276,234 +277,27 @@ pub mod legacy {
 		/// the SolSignatures needs to be converted into the RawSignature type before the
 		/// transaction is serialized as whole.
 		pub fn finalize_and_serialize(self) -> Result<Vec<u8>, bincode::error::EncodeError> {
-			bincode::serde::encode_to_vec(RawTransaction::from(self), bincode::config::legacy())
+			bincode::serde::encode_to_vec(
+				LegacyRawTransaction::from(self),
+				bincode::config::legacy(),
+			)
 		}
 	}
 
 	/// Internal raw transaction type used for correct Serialization and Encoding
 	#[derive(Debug, PartialEq, Default, Eq, Clone, Serialize, Deserialize)]
-	struct RawTransaction {
+	struct LegacyRawTransaction {
 		#[serde(with = "short_vec")]
 		pub signatures: Vec<RawSignature>,
-		pub message: Message,
+		pub message: LegacyMessage,
 	}
 
-	impl From<Transaction> for RawTransaction {
-		fn from(from: Transaction) -> Self {
+	impl From<LegacyTransaction> for LegacyRawTransaction {
+		fn from(from: LegacyTransaction) -> Self {
 			Self {
 				signatures: from.signatures.into_iter().map(RawSignature::from).collect(),
 				message: from.message,
 			}
 		}
 	}
-}
-
-/// Reasons a transaction might be rejected.
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-#[cfg_attr(feature = "std", derive(thiserror::Error))]
-pub enum TransactionError {
-	/// An account is already being processed in another transaction in a way
-	/// that does not support parallelism
-	#[cfg_attr(feature = "std", error("Account in use"))]
-	AccountInUse,
-
-	/// A `SolPubkey` appears twice in the transaction's `account_keys`.  Instructions can
-	/// reference `SolPubkey`s more than once but the message must contain a list with no duplicate
-	/// keys
-	#[cfg_attr(feature = "std", error("Account loaded twice"))]
-	AccountLoadedTwice,
-
-	/// Attempt to debit an account but found no record of a prior credit.
-	#[cfg_attr(
-		feature = "std",
-		error("Attempt to debit an account but found no record of a prior credit.")
-	)]
-	AccountNotFound,
-
-	/// Attempt to load a program that does not exist
-	#[cfg_attr(feature = "std", error("Attempt to load a program that does not exist"))]
-	ProgramAccountNotFound,
-
-	/// The from `SolPubkey` does not have sufficient balance to pay the fee to schedule the
-	/// transaction
-	#[cfg_attr(feature = "std", error("Insufficient funds for fee"))]
-	InsufficientFundsForFee,
-
-	/// This account may not be used to pay transaction fees
-	#[cfg_attr(feature = "std", error("This account may not be used to pay transaction fees"))]
-	InvalidAccountForFee,
-
-	/// The bank has seen this transaction before. This can occur under normal operation
-	/// when a UDP packet is duplicated, as a user error from a client not updating
-	/// its `recent_blockhash`, or as a double-spend attack.
-	#[cfg_attr(feature = "std", error("This transaction has already been processed"))]
-	AlreadyProcessed,
-
-	/// The bank has not seen the given `recent_blockhash` or the transaction is too old and
-	/// the `recent_blockhash` has been discarded.
-	#[cfg_attr(feature = "std", error("Blockhash not found"))]
-	BlockhashNotFound,
-
-	/// An error occurred while processing an instruction. The first element of the tuple
-	/// indicates the instruction index in which the error occurred.
-	#[cfg_attr(feature = "std", error("Error processing Instruction {0}: {1}"))]
-	InstructionError(u8, InstructionError),
-
-	/// Loader call chain is too deep
-	#[cfg_attr(feature = "std", error("Loader call chain is too deep"))]
-	CallChainTooDeep,
-
-	/// Transaction requires a fee but has no signature present
-	#[cfg_attr(feature = "std", error("Transaction requires a fee but has no signature present"))]
-	MissingSignatureForFee,
-
-	/// Transaction contains an invalid account reference
-	#[cfg_attr(feature = "std", error("Transaction contains an invalid account reference"))]
-	InvalidAccountIndex,
-
-	/// Transaction did not pass signature verification
-	#[cfg_attr(feature = "std", error("Transaction did not pass signature verification"))]
-	SignatureFailure,
-
-	/// This program may not be used for executing instructions
-	#[cfg_attr(feature = "std", error("This program may not be used for executing instructions"))]
-	InvalidProgramForExecution,
-
-	/// Transaction failed to sanitize accounts offsets correctly
-	/// implies that account locks are not taken for this TX, and should
-	/// not be unlocked.
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction failed to sanitize accounts offsets correctly")
-	)]
-	SanitizeFailure,
-
-	#[cfg_attr(
-		feature = "std",
-		error("Transactions are currently disabled due to cluster maintenance")
-	)]
-	ClusterMaintenance,
-
-	/// Transaction processing left an account with an outstanding borrowed reference
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction processing left an account with an outstanding borrowed reference")
-	)]
-	AccountBorrowOutstanding,
-
-	/// Transaction would exceed max Block Cost Limit
-	#[cfg_attr(feature = "std", error("Transaction would exceed max Block Cost Limit"))]
-	WouldExceedMaxBlockCostLimit,
-
-	/// Transaction version is unsupported
-	#[cfg_attr(feature = "std", error("Transaction version is unsupported"))]
-	UnsupportedVersion,
-
-	/// Transaction loads a writable account that cannot be written
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction loads a writable account that cannot be written")
-	)]
-	InvalidWritableAccount,
-
-	/// Transaction would exceed max account limit within the block
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction would exceed max account limit within the block")
-	)]
-	WouldExceedMaxAccountCostLimit,
-
-	/// Transaction would exceed account data limit within the block
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction would exceed account data limit within the block")
-	)]
-	WouldExceedAccountDataBlockLimit,
-
-	/// Transaction locked too many accounts
-	#[cfg_attr(feature = "std", error("Transaction locked too many accounts"))]
-	TooManyAccountLocks,
-
-	/// Address lookup table not found
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction loads an address table account that doesn't exist")
-	)]
-	AddressLookupTableNotFound,
-
-	/// Attempted to lookup addresses from an account owned by the wrong program
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction loads an address table account with an invalid owner")
-	)]
-	InvalidAddressLookupTableOwner,
-
-	/// Attempted to lookup addresses from an invalid account
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction loads an address table account with invalid data")
-	)]
-	InvalidAddressLookupTableData,
-
-	/// Address table lookup uses an invalid index
-	#[cfg_attr(feature = "std", error("Transaction address table lookup uses an invalid index"))]
-	InvalidAddressLookupTableIndex,
-
-	/// Transaction leaves an account with a lower balance than rent-exempt minimum
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction leaves an account with a lower balance than rent-exempt minimum")
-	)]
-	InvalidRentPayingAccount,
-
-	/// Transaction would exceed max Vote Cost Limit
-	#[cfg_attr(feature = "std", error("Transaction would exceed max Vote Cost Limit"))]
-	WouldExceedMaxVoteCostLimit,
-
-	/// Transaction would exceed total account data limit
-	#[cfg_attr(feature = "std", error("Transaction would exceed total account data limit"))]
-	WouldExceedAccountDataTotalLimit,
-
-	/// Transaction contains a duplicate instruction that is not allowed
-	#[cfg_attr(
-		feature = "std",
-		error("Transaction contains a duplicate instruction ({0}) that is not allowed")
-	)]
-	DuplicateInstruction(u8),
-
-	/// Transaction results in an account with insufficient funds for rent
-	#[cfg_attr(
-		feature = "std",
-		error(
-			"Transaction results in an account ({account_index}) with insufficient funds for rent"
-		)
-	)]
-	InsufficientFundsForRent { account_index: u8 },
-
-	/// Transaction exceeded max loaded accounts data size cap
-	#[cfg_attr(feature = "std", error("Transaction exceeded max loaded accounts data size cap"))]
-	MaxLoadedAccountsDataSizeExceeded,
-
-	/// LoadedAccountsDataSizeLimit set for transaction must be greater than 0.
-	#[cfg_attr(
-		feature = "std",
-		error("LoadedAccountsDataSizeLimit set for transaction must be greater than 0.")
-	)]
-	InvalidLoadedAccountsDataSizeLimit,
-
-	/// Sanitized transaction differed before/after feature activation. Needs to be resanitized.
-	#[cfg_attr(feature = "std", error("ResanitizationNeeded"))]
-	ResanitizationNeeded,
-
-	/// Program execution is temporarily restricted on an account.
-	#[cfg_attr(feature = "std", error("Execution of the program referenced by account at index {account_index} is temporarily restricted."))]
-	ProgramExecutionTemporarilyRestricted { account_index: u8 },
-
-	/// The total balance before the transaction does not equal the total balance after the
-	/// transaction
-	#[cfg_attr(
-		feature = "std",
-		error("Sum of account balances before and after transaction do not match")
-	)]
-	UnbalancedTransaction,
 }
